src/data_ingestion.py
import os
import pandas as pd
from google.cloud import storage
from sklearn.model_selection import train_test_split
from src.logger import get_logger
from src.custom_exception import CustomException
from config.path_config import *   # Import all path defined in path_config.py
from utils.common_functions import read_yaml 

logger = get_logger(__name__) 


class DataIngestion:
    def __init__(self, config):
        self.config = config["data_ingestion"]   
        self.bucket_name = self.config['bucket_name'] 
        self.bucket_file = self.config['bucket_file_name']
        self.train_test_ratio = self.config["train_ratio"]

        os.makedirs(RAW_DIR, exist_ok= True)

        logger.info(f"Data ingestion started with {self.bucket_name} and file is {self.bucket_file}")


    def download_csv_from_gcp(self):
        try:
            client = storage.Client()
            bucket= client.bucket(self.bucket_name)
            blob = bucket.blob(self.bucket_file) 

            blob.download_to_filename(RAW_FILE_PATH)

            logger.info(f"Raw file successfully downloaded to {RAW_FILE_PATH}")

        except Exception as e:
            logger.error("Error while downloading the csv file")
            raise CustomException("Failed to download csv file", e)
        

    def split_data(self):
        try:
            logger.info("Starting the splitting process...")
            data = pd.read_csv(RAW_FILE_PATH)  #The file went to raw file path from GCP
            
            train_data, test_data = train_test_split(data, test_size= 1-self.train_test_ratio, random_state= 42)
            
            # Converting from DataFrame to csv
            train_data.to_csv(TRAIN_FILE_PATH)
            test_data.to_csv(TEST_FILE_PATH)

            logger.info(f"Train data saved to {TRAIN_FILE_PATH}")
            logger.info(f"Test data saved to {TEST_FILE_PATH}")

        except Exception as e:
            logger.error("Error while splitting data")
            raise CustomException("Failed to split data into training and test sets", e)


    def run(self):
        """
        W/o this, we'll create data ingection method with --
            obj = DataIngestion()
            obj.download_csv_from_gcp()
            obj.split_data

        This function combine the 3 steps into 1 neat step
        """
        try:
            logger.info(f"Starting data ingestion process...")
            
            self.download_csv_from_gcp()
            self.split_data()

            logger.info("Data ingestion completed succesfully")

        except CustomException as ce:
            logger.error(f"CustomException: {str(ce)}")
        
        finally:  # Will be executed regardless of success or not
            logger.info("Data ingestion completed")


if __name__ == "__main__":  # Content under this lane will be executed whenever we run data_ingestion.py
    data_ingestion = DataIngestion(read_yaml(CONFIG_PATH))
    data_ingestion.run()
